rrr: convert rpc/fiber_channel to module (rrr.fiber_channel)
Closes the rrr module migration at 48/48. fiber_channel.cpp becomes
a C++23 named-module interface unit; fiber_channel.hpp shrinks to a
6-line anchor shim that `#include`s `<memory>`. `rrr/rrr.hpp` keeps
including that shim and adds `import rrr.fiber_channel;` to its
import list. `rrr/rpc/client.cpp` drops the textual
`#include "fiber_channel.hpp"` (the import covers it).

Diagnosis of the bug that blocked this conversion in the prior
attempt (recorded in docs/dev/srpc_module_migration_plan.md):
importing `rrr.fiber_channel` into a TU that also `import std;`s
makes clang 22 ambiguate `operator new(size_t, std::align_val_t)`
inside `std::__libcpp_allocate<std::shared_ptr<rusty::Waker>>`
(and `<rusty::Task<void>>`) instantiations downstream — typically in
`rpcbench.cc` via the umbrella. Other rpc modules avoid the
ambiguity for reasons we don't fully understand.

Workaround (empirical): keep a tiny anchor shim header
`fiber_channel.hpp` with `#include <memory>` and have `rrr/rrr.hpp`
`#include` it *before* `import rrr.fiber_channel;`. The textual
`<memory>` pins libc++ `operator new` in the global module ahead of
the import, and the ambiguity vanishes. No other change to the
module body is required. Tested clean build of rrr + rpcbench under
clang 22.1.5 in 77 s.

The other clang 22 codegen quirk that blocks fresh libc++ container
template instantiations from module purview is sidestepped via the
standard `#include <rusty/rusty.hpp>` GMF trick (same as alock,
tcp_channel, etc.).
